+++ src/components/content/timeline/patient-timeline.tsx
@@ -8,6 +8,7 @@ import { useObservationsDetailsDrawer } from "../observations/helpers/drawer";
 import { getDateRangeView } from "../resource/helpers/view-date-range";
 import { ResourceTableActions } from "../resource/resource-table-actions";
 import { ResourceTable } from "@/components/content/resource/resource-table";
+import { EmptyTable } from "@/components/core/empty-table";
 import { TimelineEventModel } from "@/fhir/models/timeline-event";
 import { useTimelineEvents } from "@/fhir/timeline-event";
 import { useFilteredSortedData } from "@/hooks/use-filtered-sorted-data";
@@ -26,6 +27,10 @@ export function PatientTimeline({ className }: PatientTimelineProps) {
     defaultSort: defaultTimelineSort,
     records: timelineEventsQuery.data,
   });
+
+  const isEmptyQuery = timelineEventsQuery.data.length === 0;
+  const hasZeroFilteredRecords = !isEmptyQuery && data.length === 0;
+
   const openEncounterDetails = usePatientEncounterDetailsDrawer();
   const openDiagnosticReportDetails = useObservationsDetailsDrawer();
 
@@ -52,7 +57,12 @@ export function PatientTimeline({ className }: PatientTimelineProps) {
         showTableHead={false}
         isLoading={timelineEventsQuery.isLoading}
         data={data}
-        emptyMessage="There are no timeline records available."
+        emptyMessage={
+          <EmptyTable
+            hasZeroFilteredRecords={hasZeroFilteredRecords}
+            resourceName="timeline records"
+          />
+        }
         columns={patientTimelineColumns}
         onRowClick={(record) => {
           switch (record.model.kind) {